requirements.txt
python-telegram-bot==21.3
openai==1.40.0
